move db keys back to common, move funcs to state
